refactor: Optimizing imports
Move internal mappings logic to own subfolder
