+++ src/chunked_data_view/ViewPart.cc
@@ -148,24 +148,19 @@ ViewPart::ViewPart(const metkit::mars::MarsRequest& request, const DataLayout& d
     extension_.reserve(axes_.size());
     std::transform(std::begin(axes_), std::end(axes_), std::back_inserter(extension_),
                    [](const auto& axis) { return axis.size(); });
-    // extension_.push_back(data_layout.countValues);
 
-
-    auto lower = offset_;
-    // lower.push_back(0);  // Add dimensions for the implicit values
-
+    const auto lower = offset_;
     auto upper = offset_;
 
     for (size_t i = 0; i < offset_.size(); ++i) {
         upper[i] += (extension_[i] - 1);
     }
-    // upper.push_back(data_layout.countValues - 1);  // Add dimensions for the implicit values
 
     bb_ = BoundingBox(lower, upper);
 }
 
 
-metkit::mars::MarsRequest ViewPart::at(const ChunkedDataViewPartBoundingBox& boundingBox) const {
+metkit::mars::MarsRequest ViewPart::at(const PartBoundingBox& boundingBox) const {
 
     const auto translatedBB = bb_.subtract(bb_.lower());
     const auto intersection = translatedBB.intersect(boundingBox);
