+++ src/chunked_data_view/include/chunked_data_view/ChunkedDataViewBuilder.h
@@ -26,28 +26,51 @@ namespace chunked_data_view {
 
 class ViewPart;
 
+/// Fluent builder for constructing a ChunkedDataView from one or more MARS data parts.
+///
+/// Usage pattern:
+///   1. Construct the builder (optionally supplying an FDB config path).
+///   2. Call addPart() once per logical data region (e.g. once for surface fields, once for
+///      pressure-level fields).
+///   3. If more than one part is added, call extendOnAxis() to specify which axis stitches the
+///      parts together.
+///   4. Optionally call fillValue() to set the sentinel for grid positions with no data.
+///   5. Call build() to validate and assemble the ChunkedDataView.
 class ChunkedDataViewBuilder {
 public:
 
+    /// @param configPath  Optional path to an FDB config file. Passed through to the FDB
+    ///                    instance created when building with ExtractorType::GRIB.
     explicit ChunkedDataViewBuilder(const std::optional<std::filesystem::path>& configPath = std::nullopt);
-    /// Add data to the view defined by a mars request.
-    /// Multiple parts can be added by repeated calls to addPart, e.g. adding surface and non surface fields.
-    /// Every keyword with more than one value needs to be mapped to the output axes with an AxisDefinition.
-    /// Multiple keywords can map to one axis.
+
+    /// Registers one data region (part) of the view.
+    ///
+    /// Each keyword in @p marsRequestKeyValues that has more than one value must appear in
+    /// exactly one AxisDefinition in @p axes. Multiple keywords may share one axis
+    /// (they form a compound Cartesian-product axis).
+    ///
+    /// Multiple parts can cover different variable types (e.g. surface and pressure-level
+    /// fields) and are stitched together along the extension axis specified by extendOnAxis().
     ChunkedDataViewBuilder& addPart(std::string marsRequestKeyValues, std::vector<AxisDefinition> axes,
                                     std::shared_ptr<Extractor> extractor);
 
-    /// On which axis the multiple parts extend each other.
-    /// For multiple part builds this all other axis must have the same extension. For a single
-    /// part build this index is simply ignored.
+    /// Sets the axis index along which multiple parts are concatenated.
     ///
-    /// @throws eckit::UserException in case the index is not a valid axis
+    /// All parts must have identical extents on every axis except this one.
+    /// For a single-part view this call is optional and the value is ignored.
+    ///
+    /// @throws eckit::UserError if @p index exceeds the number of axes in the first part.
     ChunkedDataViewBuilder& extendOnAxis(size_t index);
 
+    /// Sets the fill value written to array positions not covered by any part.
+    /// Defaults to positive infinity.
     ChunkedDataViewBuilder& fillValue(float fillValue);
 
+    /// Validates all parts, checks axis compatibility, and returns the assembled view.
+    /// @throws eckit::UserError on misconfiguration (missing parts, axis mismatch, etc.).
     std::unique_ptr<ChunkedDataView> build();
 
+    /// Returns the FDB config path supplied at construction, if any.
     std::optional<std::filesystem::path> getFdbConfigPath() const { return configPath_; }
 
 private:
