+++ src/chunked_data_view/include/chunked_data_view/DataLayout.h
@@ -13,9 +13,11 @@
 
 namespace chunked_data_view {
 
+/// Describes the binary layout of a single GRIB field's value array.
+/// All fields within one ViewPart are required to share the same layout.
 struct DataLayout {
-    size_t countValues{};
-    size_t bytesPerValue{};
+    size_t countValues{};   ///< Number of floating-point values in the field (e.g. grid points).
+    size_t bytesPerValue{}; ///< Storage size of each value in bytes (typically 4 for float32).
 };
 
 }  // namespace chunked_data_view
