+++ src/chunked_data_view/include/chunked_data_view/Fdb.h
@@ -21,14 +21,26 @@
 
 namespace chunked_data_view {
 
+/// Abstract interface over an FDB instance, allowing the real FDB to be swapped out for
+/// a mock in tests.
+///
+/// Two operations are needed: retrieving the raw field bytes for a single field (used to
+/// probe the DataLayout), and listing all matching fields with their keys and data handles
+/// (used when filling a chunk).
 class FdbInterface {
 public:
 
     virtual ~FdbInterface() = default;
+
+    /// Returns a DataHandle positioned at the start of the field data for @p request.
+    /// Used by Extractor::layout() to read the DataLayout of a single representative field.
     virtual std::unique_ptr<eckit::DataHandle> retrieve(const metkit::mars::MarsRequest& request) = 0;
+
+    /// Returns an iterator over all fields matching @p request, yielding (key, data handle) pairs.
     virtual std::unique_ptr<ListIteratorInterface> inspect(const metkit::mars::MarsRequest& request) = 0;
 };
 
+/// Creates a real FDB instance, optionally configured from @p configPath.
 std::unique_ptr<FdbInterface> makeFdb(std::optional<std::filesystem::path> configPath = std::nullopt);
 
 };  // namespace chunked_data_view
