+++ src/chunked_data_view/include/chunked_data_view/ListIterator.h
@@ -21,15 +21,22 @@
 
 namespace chunked_data_view {
 
+/// Abstract iterator over FDB fields matching a MARS request.
+///
+/// Each call to next() yields the MARS key and a data handle for one matching field,
+/// or std::nullopt when the sequence is exhausted.
 class ListIteratorInterface {
 
 public:
 
     virtual ~ListIteratorInterface() = default;
+
+    /// Returns the next (key, data-handle) pair, or std::nullopt if there are no more fields.
     virtual std::optional<std::tuple<fdb5::Key, std::unique_ptr<eckit::DataHandle>>> next() = 0;
 };
 
 
+/// Wraps a real fdb5::ListIterator to satisfy the ListIteratorInterface contract.
 class ListIteratorWrapperImpl : public ListIteratorInterface {
 
     fdb5::ListIterator listIterator_;
@@ -40,5 +47,6 @@ public:
     std::optional<std::tuple<fdb5::Key, std::unique_ptr<eckit::DataHandle>>> next() override;
 };
 
+/// Wraps @p listIterator in a ListIteratorInterface-compatible heap object.
 std::unique_ptr<ListIteratorInterface> makeListIterator(fdb5::ListIterator listIterator);
 };  // namespace chunked_data_view
