+++ src/chunked_data_view/include/chunked_data_view/exception/GribExtractorException.h
@@ -12,6 +12,8 @@
 
 namespace chunked_data_view {
 
+/// Thrown by GribExtractor when field retrieval fails (e.g. the FDB iterator returns no
+/// results for a request, or a field's value count does not match the expected DataLayout).
 class GribExtractorException : public eckit::Exception {
 
 public:
