+++ src/chunked_data_view/include/chunked_data_view/exception/RequestManipulationException.h
@@ -12,6 +12,8 @@
 
 namespace chunked_data_view {
 
+/// Thrown by RequestManipulation when a bounding box cannot be mapped to a single MARS request
+/// (e.g. the box is not aligned with a sub-axis boundary, which would require multiple requests).
 class RequestManipulationException : public eckit::Exception {
 
 public:
