+++ src/chunked_data_view/include/chunked_data_view/exception/UnknownExtractorException.h
@@ -11,6 +11,7 @@
 
 namespace chunked_data_view {
 
+/// Thrown when an unrecognised ExtractorType is supplied to the factory.
 class UnknownExtractorException : public eckit::Exception {
 
 public:
