+++ tests/chunked_data_view/test_index_mapper.cc
@@ -2,7 +2,7 @@
 #include <string>
 #include <vector>
 #include "chunked_data_view/Axis.h"
-#include "chunked_data_view/IndexMapper.h"
+#include "chunked_data_view/mapping/IndexMapper.h"
 #include "eckit/testing/Test.h"
 #include "fdb5/database/Key.h"
